feat: Add core components and hooks for website functionality
- Implemented FloatingCTA component for a prominent call-to-action button.
- Created Footer and Header components for site navigation and language selection.
- Developed HeroSection for the landing page with animated headlines.
- Added LazyWrapper for lazy loading of components.
- Built ProcessSection to display steps in a user-friendly manner.
- Created ServicesSection to showcase various services offered.
- Developed ValuePropositions component to highlight key offerings.
- Implemented custom hooks: useContactForm for form handling, useIntersectionObserver for scroll tracking, useLanguage for language management, and useScrollTracking for scroll state detection.
- Added performance utilities for measuring function execution time and debouncing/throttling functions.
- Created service configuration and translation files for multi-language support.
- Defined types for translations, services, and form data structures.
