+++ components/LazyWrapper.tsx
@@ -1,0 +1,17 @@
+import React, { Suspense } from 'react'
+
+interface LazyWrapperProps {
+  children: React.ReactNode
+  fallback?: React.ReactNode
+}
+
+export const LazyWrapper: React.FC<LazyWrapperProps> = ({ 
+  children, 
+  fallback = <div className="animate-pulse bg-gray-200 rounded h-32 w-full" /> 
+}) => {
+  return (
+    <Suspense fallback={fallback}>
+      {children}
+    </Suspense>
+  )
+}
