+++ hooks/useIntersectionObserver.ts
@@ -1,0 +1,54 @@
+import { useEffect, useRef, useState } from 'react'
+
+export const useIntersectionObserver = (threshold: number = 0.1) => {
+  const [isInView, setIsInView] = useState(false)
+  const [hasBeenInView, setHasBeenInView] = useState(false)
+  const ref = useRef<HTMLDivElement>(null)
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry.isIntersecting) {
+          setIsInView(true)
+          setHasBeenInView(true)
+        } else {
+          setIsInView(false)
+        }
+      },
+      { threshold }
+    )
+
+    if (ref.current) {
+      observer.observe(ref.current)
+    }
+
+    return () => observer.disconnect()
+  }, [threshold])
+
+  return { ref, isInView, hasBeenInView }
+}
+
+export const useOnceInView = (threshold: number = 0.1) => {
+  const [isInView, setIsInView] = useState(false)
+  const ref = useRef<HTMLDivElement>(null)
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry.isIntersecting && !isInView) {
+          setIsInView(true)
+          observer.disconnect()
+        }
+      },
+      { threshold }
+    )
+
+    if (ref.current) {
+      observer.observe(ref.current)
+    }
+
+    return () => observer.disconnect()
+  }, [threshold, isInView])
+
+  return { ref, isInView }
+}
