+++ lib/performance.ts
@@ -1,0 +1,50 @@
+/**
+ * Performance monitoring utilities
+ */
+
+export const measurePerformance = <T>(
+  name: string,
+  fn: () => T
+): T => {
+  const start = performance.now()
+  const result = fn()
+  const end = performance.now()
+  console.log(`${name} took ${end - start} milliseconds`)
+  return result
+}
+
+export const measureAsyncPerformance = async <T>(
+  name: string,
+  fn: () => Promise<T>
+): Promise<T> => {
+  const start = performance.now()
+  const result = await fn()
+  const end = performance.now()
+  console.log(`${name} took ${end - start} milliseconds`)
+  return result
+}
+
+export const debounce = <T extends (...args: any[]) => void>(
+  func: T,
+  delay: number
+): T => {
+  let timeoutId: NodeJS.Timeout
+  return ((...args: Parameters<T>) => {
+    clearTimeout(timeoutId)
+    timeoutId = setTimeout(() => func(...args), delay)
+  }) as T
+}
+
+export const throttle = <T extends (...args: any[]) => void>(
+  func: T,
+  limit: number
+): T => {
+  let inThrottle: boolean
+  return ((...args: Parameters<T>) => {
+    if (!inThrottle) {
+      func(...args)
+      inThrottle = true
+      setTimeout(() => (inThrottle = false), limit)
+    }
+  }) as T
+}
