+++ lib/services.ts
@@ -1,0 +1,45 @@
+import {
+  Search,
+  Handshake,
+  FileText,
+  MapPin,
+  Megaphone,
+  Building,
+  Calculator,
+  Network,
+} from "lucide-react"
+
+export const serviceConfig = [
+  {
+    id: 'service1',
+    icon: Search,
+  },
+  {
+    id: 'service2',
+    icon: Handshake,
+  },
+  {
+    id: 'service3',
+    icon: FileText,
+  },
+  {
+    id: 'service4',
+    icon: MapPin,
+  },
+  {
+    id: 'service5',
+    icon: Megaphone,
+  },
+  {
+    id: 'service6',
+    icon: Building,
+  },
+  {
+    id: 'service7',
+    icon: Calculator,
+  },
+  {
+    id: 'service8',
+    icon: Network,
+  },
+]
