+++ lib/translations/additional.ts
@@ -1,0 +1,62 @@
+import { Translations } from '../types'
+
+export const additionalTranslations: Translations = {
+  // Process Section
+  ourProcess: { en: "Our Process", lt: "Mūsų procesas", no: "Vår prosess" },
+  step1: { en: "Strategy Call", lt: "Strategijos skambutis", no: "Strategisamtale" },
+  step1Desc: {
+    en: "Understanding your goals and market ambitions",
+    lt: "Jūsų tikslų ir rinkos ambicijų supratimas",
+    no: "Forstå dine mål og markedsambisjoner",
+  },
+  step2: { en: "Market Entry Blueprint", lt: "Rinkos patekimo planas", no: "Markedsinntreden-plan" },
+  step2Desc: {
+    en: "Detailed roadmap with timelines and milestones",
+    lt: "Detalus kelių žemėlapis su terminais ir etapais",
+    no: "Detaljert veikart med tidslinjer og milepæler",
+  },
+  step3: { en: "Execution Phase", lt: "Įgyvendinimo etapas", no: "Gjennomføringsfase" },
+  step3Desc: {
+    en: "Hands-on implementation of your expansion strategy",
+    lt: "Praktinis jūsų plėtros strategijos įgyvendinimas",
+    no: "Praktisk implementering av din ekspansjonsstrategi",
+  },
+  step4: { en: "Ongoing Support", lt: "Nuolatinis palaikymas", no: "Kontinuerlig støtte" },
+  step4Desc: {
+    en: "Continuous optimization and growth support",
+    lt: "Nuolatinis optimizavimas ir augimo palaikymas",
+    no: "Kontinuerlig optimalisering og vekststøtte",
+  },
+
+  // Contact Section
+  readyToGrow: {
+    en: "Ready to grow across the region?",
+    lt: "Pasiruošę augti visame regione?",
+    no: "Klar til å vokse på tvers av regionen?",
+  },
+  fullName: { en: "Full Name", lt: "Pilnas vardas", no: "Fullt navn" },
+  company: { en: "Company", lt: "Įmonė", no: "Selskap" },
+  email: { en: "Email", lt: "El. paštas", no: "E-post" },
+  interestedIn: { en: "I'm interested in:", lt: "Man įdomu:", no: "Jeg er interessert i:" },
+  balticNordic: {
+    en: "Baltic → Nordic expansion",
+    lt: "Baltijos → Šiaurės šalių plėtra",
+    no: "Baltisk → Nordisk ekspansjon",
+  },
+  nordicBaltic: {
+    en: "Nordic → Baltic expansion",
+    lt: "Šiaurės šalių → Baltijos plėtra",
+    no: "Nordisk → Baltisk ekspansjon",
+  },
+  message: { en: "Message", lt: "Žinutė", no: "Melding" },
+  sendMessage: { en: "Send Message", lt: "Siųsti žinutę", no: "Send melding" },
+
+  // Footer
+  footer: {
+    en: "© 2025 ExpandNord. All rights reserved.",
+    lt: "© 2025 ExpandNord. Visos teisės saugomos.",
+    no: "© 2025 ExpandNord. Alle rettigheter forbeholdt.",
+  },
+  privacy: { en: "Privacy", lt: "Privatumas", no: "Personvern" },
+  terms: { en: "Terms", lt: "Sąlygos", no: "Vilkår" },
+}
