+++ lib/translations/core.ts
@@ -1,0 +1,84 @@
+import { Translations } from '../types'
+
+export const coreTranslations: Translations = {
+  // Navigation
+  about: { en: "About", lt: "Apie mus", no: "Om oss" },
+  services: { en: "Services", lt: "Paslaugos", no: "Tjenester" },
+  process: { en: "Process", lt: "Procesas", no: "Prosess" },
+  contact: { en: "Contact", lt: "Kontaktai", no: "Kontakt" },
+  bookCall: { en: "Book Strategy Call", lt: "Susisiekime", no: "Bestill strategisamtale" },
+
+  // Hero Section
+  heroTitle: {
+    en: "ExpandNord – Growth by Connecting Baltic–Nordic Business",
+    lt: "ExpandNord – Aukite Šiaurės šalyse",
+    no: "ExpandNord – Vekst gjennom å koble baltisk-nordisk næringsliv",
+  },
+  heroSubtitle: {
+    en: "We bridge markets, cultures, and opportunities between the Baltic and Nordic regions, turning expansion challenges into growth success stories.",
+    lt: "Mes sujungiame rinkas, kultūras ir galimybes tarp Baltijos ir Šiaurės šalių regionų, paversdami plėtros iššūkius augimo sėkmės istorijomis.",
+    no: "Vi bygger broer mellom markeder, kulturer og muligheter mellom de baltiske og nordiske regionene, og gjør ekspansjonsutfordringer til vekstsuksesshistorier.",
+  },
+  getStarted: { en: "Get Started Today", lt: "Pradėkite šiandien", no: "Kom i gang i dag" },
+
+  // Value Bullets
+  marketIntel: { en: "Market Intelligence", lt: "Rinkos analizė", no: "Markedsintelligens" },
+  marketIntelDesc: {
+    en: "Deep insights into local markets and opportunities",
+    lt: "Gilūs vietinių rinkų ir galimybių įžvalgos",
+    no: "Dype innsikter i lokale markeder og muligheter",
+  },
+  connections: {
+    en: "Connections That Open Doors",
+    lt: "Ryšiai, kurie atveria duris",
+    no: "Forbindelser som åpner dører",
+  },
+  connectionsDesc: {
+    en: "Access to key decision-makers and industry leaders",
+    lt: "Prieiga prie pagrindinių sprendimų priėmėjų ir pramonės lyderių",
+    no: "Tilgang til sentrale beslutningstakere og bransjeledere",
+  },
+  strategy: {
+    en: "Strategy That Leads to Sales",
+    lt: "Strategija, vedanti į pardavimus",
+    no: "Strategi som fører til salg",
+  },
+  strategyDesc: {
+    en: "Proven methodologies that convert expansion into revenue",
+    lt: "Patirtis, kurianti pajamas",
+    no: "Dokumenterte metoder som konverterer ekspansjon til inntekter",
+  },
+
+  // About Section
+  ourMission: { en: "Our Mission", lt: "Mūsų misija", no: "Vårt oppdrag" },
+  missionText: {
+    en: "We turn ambition into opportunity between Baltic and Nordic markets.",
+    lt: "Mes paverčiame ambicijas galimybėmis tarp Baltijos ir Šiaurės šalių rinkų.",
+    no: "Vi forvandler ambisjoner til muligheter mellom baltiske og nordiske markeder.",
+  },
+  whyWorkWithUs: {
+    en: "Why Companies Work With Us:",
+    lt: "Kodėl įmonės dirba su mumis:",
+    no: "Hvorfor selskaper jobber med oss:",
+  },
+  reason1: {
+    en: "Local expertise in both Baltic and Nordic markets",
+    lt: "Vietinė ekspertizė tiek Baltijos, tiek Šiaurės šalių rinkose",
+    no: "Lokal ekspertise i både baltiske og nordiske markeder",
+  },
+  reason2: {
+    en: "Established network of partners and decision-makers",
+    lt: "Sukurtas partnerių ir sprendimų priėmėjų tinklas",
+    no: "Etablert nettverk av partnere og beslutningstakere",
+  },
+  reason3: {
+    en: "End-to-end support from strategy to execution",
+    lt: "Visapusiškas palaikymas nuo strategijos iki įgyvendinimo",
+    no: "Ende-til-ende støtte fra strategi til gjennomføring",
+  },
+  reason4: {
+    en: "Proven track record of successful market entries",
+    lt: "Įrodytas sėkmingų rinkos patekimų rekordas",
+    no: "Dokumentert merittliste med vellykkede markedsinntreden",
+  },
+}
