+++ lib/translations/services.ts
@@ -1,0 +1,164 @@
+import { ServiceTranslation } from '../types'
+
+export const serviceTranslations: Record<string, ServiceTranslation> = {
+  service1: {
+    title: {
+      en: "Market Research & Strategy",
+      lt: "Rinkos tyrimai ir strategija",
+      no: "Markedsundersøkelse og strategi",
+    },
+    points: {
+      en: ["Comprehensive market analysis", "Competitive landscape mapping", "Entry strategy development"],
+      lt: ["Išsami rinkos analizė", "Konkurencinės aplinkos vertinimas", "Patekimo strategijos kūrimas"],
+      no: ["Omfattende markedsanalyse", "Kartlegging av konkurranselandskap", "Utvikling av inntrengningsstrategi"],
+    },
+    closing: {
+      en: "Data-driven insights for confident decisions",
+      lt: "Duomenimis grįsti sprendimai",
+      no: "Datadrevne innsikter for trygge beslutninger",
+    },
+  },
+  service2: {
+    title: { en: "B2B & B2G Meetings", lt: "B2B ir B2G susitikimai", no: "B2B og B2G møter" },
+    points: {
+      en: ["Pre-qualified prospect meetings", "Government liaison introductions", "Industry event coordination"],
+      lt: ["Susitikimai su potencialiais partneriais", "Ryšiai su valdžios institucijomis", "Renginių koordinavimas"],
+      no: [
+        "Forhåndskvalifiserte prospektmøter",
+        "Introduksjoner til myndighetskontakter",
+        "Koordinering av bransjearrangementer",
+      ],
+    },
+    closing: {
+      en: "Direct access to decision-makers",
+      lt: "Prieiga prie sprendimų priėmėjų",
+      no: "Direkte tilgang til beslutningstakere",
+    },
+  },
+  service3: {
+    title: {
+      en: "Registration & Legal Setup",
+      lt: "Įmonės registracija",
+      no: "Registrering og juridisk oppsett",
+    },
+    points: {
+      en: ["Company registration assistance", "Legal structure optimization", "Compliance framework setup"],
+      lt: ["Pagalba registruojant įmonę", "Teisinės struktūros pritaikymas", "Atitikties sistemos sąranka"],
+      no: [
+        "Assistanse med selskapsregistrering",
+        "Optimalisering av juridisk struktur",
+        "Oppsett av compliance-rammeverk",
+      ],
+    },
+    closing: {
+      en: "Legally compliant market entry",
+      lt: "Teisiškai atitinkantis rinkos patekimas",
+      no: "Juridisk kompatibel markedsinntreden",
+    },
+  },
+  service4: {
+    title: { en: "Location & Setup", lt: "Vieta ir sąranka", no: "Lokasjon og oppsett" },
+    points: {
+      en: ["Office space identification", "Virtual office solutions", "Operational infrastructure setup"],
+      lt: ["Biuro erdvės identifikavimas", "Virtualaus biuro sprendimai", "Veiklos infrastruktūros sąranka"],
+      no: ["Identifisering av kontorlokaler", "Virtuelle kontorløsninger", "Oppsett av operasjonell infrastruktur"],
+    },
+    closing: {
+      en: "Physical presence without the hassle",
+      lt: "Fizinis buvimas be rūpesčių",
+      no: "Fysisk tilstedeværelse uten problemer",
+    },
+  },
+  service5: {
+    title: { en: "PR & Marketing", lt: "PR ir rinkodara", no: "PR og markedsføring" },
+    points: {
+      en: ["Local media relationship building", "Content localization strategy", "Digital marketing campaigns"],
+      lt: [
+        "Vietinių žiniasklaidos ryšių kūrimas",
+        "Turinio lokalizacijos strategija",
+        "Skaitmeninės rinkodaros kampanijos",
+      ],
+      no: ["Bygging av lokale medierelasjoner", "Strategi for innholdslokalisering", "Digitale markedsføringskampanjer"],
+    },
+    closing: {
+      en: "Brand visibility in new markets",
+      lt: "Prekės ženklo matomumas naujose rinkose",
+      no: "Merkevaresynlighet i nye markeder",
+    },
+  },
+  service6: {
+    title: {
+      en: "Government & Institutional Engagement",
+      lt: "Vyriausybės ir institucinė veikla",
+      no: "Myndighets- og institusjonell engasjement",
+    },
+    points: {
+      en: ["Regulatory navigation support", "Government incentive identification", "Public sector relationship building"],
+      lt: [
+        "Reguliavimo navigacijos palaikymas",
+        "Vyriausybės paskatos identifikavimas",
+        "Viešojo sektorius ryšių kūrimas",
+      ],
+      no: [
+        "Støtte til regulatorisk navigering",
+        "Identifisering av offentlige insentiver",
+        "Bygging av offentlige sektorrelasjoner",
+      ],
+    },
+    closing: {
+      en: "Leverage public sector opportunities",
+      lt: "Pasinaudokite viešojo sektorius galimybėmis",
+      no: "Utnytt offentlige sektormuligheter",
+    },
+  },
+  service7: {
+    title: {
+      en: "Tax & Financial Structuring",
+      lt: "Mokesčių ir finansinė struktūra",
+      no: "Skatte- og finansiell strukturering",
+    },
+    points: {
+      en: ["Tax optimization strategies", "Financial structure planning", "Cross-border transaction support"],
+      lt: [
+        "Mokesčių optimizavimo strategijos",
+        "Finansinės struktūros planavimas",
+        "Tarpvalstybinių sandorių palaikymas",
+      ],
+      no: [
+        "Skatteoptimaliseringsstrategier",
+        "Planlegging av finansiell struktur",
+        "Støtte til grenseoverskridende transaksjoner",
+      ],
+    },
+    closing: {
+      en: "Maximize financial efficiency",
+      lt: "Maksimalus finansinis efektyvumas",
+      no: "Maksimer finansiell effektivitet",
+    },
+  },
+  service8: {
+    title: {
+      en: "Partner & Distributor Search",
+      lt: "Partnerių ir platintojų paieška",
+      no: "Partner- og distributørsøk",
+    },
+    points: {
+      en: ["Strategic partner identification", "Distributor network mapping", "Partnership negotiation support"],
+      lt: [
+        "Strateginių partnerių identifikavimas",
+        "Platintojų tinklo žemėlapių sudarymas",
+        "Partnerystės derybų palaikymas",
+      ],
+      no: [
+        "Identifisering av strategiske partnere",
+        "Kartlegging av distributørnettverk",
+        "Støtte til partnerskapsforhandlinger",
+      ],
+    },
+    closing: {
+      en: "Accelerate market penetration",
+      lt: "Efektyvi plėtra",
+      no: "Akseler markedspenetrasjon",
+    },
+  },
+}
